heuristics to establish a tolerance band for watching global trends
After the individual tests, we calculate the averaged delta over the
whole test suite, to detect changes to the overall timings. As it turned out,
using the error propagation for the calculation of the averaged delta
yields the right tolerance band to ignore random fluctuations but
trigger alarm on real changes.

Moreover, add several further timing test cases
to verify the calibration via "platform model" works as intended